+++ zcash_primitives/src/transaction/components/amount.rs
@@ -1,3 +1,4 @@
+use std::convert::TryFrom;
 use std::iter::Sum;
 use std::ops::{Add, AddAssign, Sub, SubAssign};
 
@@ -101,6 +102,14 @@ impl Amount {
     }
 }
 
+impl TryFrom<i64> for Amount {
+    type Error = ();
+
+    fn try_from(value: i64) -> Result<Self, ()> {
+        Amount::from_i64(value)
+    }
+}
+
 impl From<Amount> for i64 {
     fn from(amount: Amount) -> i64 {
         amount.0
@@ -114,36 +123,52 @@ impl From<Amount> for u64 {
 }
 
 impl Add<Amount> for Amount {
-    type Output = Amount;
+    type Output = Option<Amount>;
 
-    fn add(self, rhs: Amount) -> Amount {
-        Amount::from_i64(self.0 + rhs.0).expect("addition should remain in range")
+    fn add(self, rhs: Amount) -> Option<Amount> {
+        Amount::from_i64(self.0 + rhs.0).ok()
+    }
+}
+
+impl Add<Amount> for Option<Amount> {
+    type Output = Self;
+
+    fn add(self, rhs: Amount) -> Option<Amount> {
+        self.and_then(|lhs| lhs + rhs)
     }
 }
 
 impl AddAssign<Amount> for Amount {
     fn add_assign(&mut self, rhs: Amount) {
-        *self = *self + rhs
+        *self = (*self + rhs).expect("Addition must produce a valid amount value.")
     }
 }
 
 impl Sub<Amount> for Amount {
-    type Output = Amount;
+    type Output = Option<Amount>;
 
-    fn sub(self, rhs: Amount) -> Amount {
-        Amount::from_i64(self.0 - rhs.0).expect("subtraction should remain in range")
+    fn sub(self, rhs: Amount) -> Option<Amount> {
+        Amount::from_i64(self.0 - rhs.0).ok()
+    }
+}
+
+impl Sub<Amount> for Option<Amount> {
+    type Output = Self;
+
+    fn sub(self, rhs: Amount) -> Option<Amount> {
+        self.and_then(|lhs| lhs - rhs)
     }
 }
 
 impl SubAssign<Amount> for Amount {
     fn sub_assign(&mut self, rhs: Amount) {
-        *self = *self - rhs
+        *self = (*self - rhs).expect("Subtraction must produce a valid amount value.")
     }
 }
 
-impl Sum for Amount {
-    fn sum<I: Iterator<Item = Amount>>(iter: I) -> Amount {
-        iter.fold(Amount::zero(), Add::add)
+impl Sum<Amount> for Option<Amount> {
+    fn sum<I: Iterator<Item = Amount>>(iter: I) -> Self {
+        iter.fold(Some(Amount::zero()), |acc, a| acc? + a)
     }
 }
 
@@ -154,7 +179,19 @@ pub mod testing {
     use super::{Amount, MAX_MONEY};
 
     prop_compose! {
+        pub fn arb_amount()(amt in -MAX_MONEY..MAX_MONEY) -> Amount {
+            Amount::from_i64(amt).unwrap()
+        }
+    }
+
+    prop_compose! {
         pub fn arb_nonnegative_amount()(amt in 0i64..MAX_MONEY) -> Amount {
+            Amount::from_i64(amt).unwrap()
+        }
+    }
+
+    prop_compose! {
+        pub fn arb_positive_amount()(amt in 1i64..MAX_MONEY) -> Amount {
             Amount::from_i64(amt).unwrap()
         }
     }
@@ -213,10 +250,9 @@ mod tests {
     }
 
     #[test]
-    #[should_panic]
-    fn add_panics_on_overflow() {
+    fn add_overflow() {
         let v = Amount(MAX_MONEY);
-        let _sum = v + Amount(1);
+        assert_eq!(v + Amount(1), None)
     }
 
     #[test]
@@ -227,10 +263,9 @@ mod tests {
     }
 
     #[test]
-    #[should_panic]
-    fn sub_panics_on_underflow() {
+    fn sub_underflow() {
         let v = Amount(-MAX_MONEY);
-        let _diff = v - Amount(1);
+        assert_eq!(v - Amount(1), None)
     }
 
     #[test]